curvefs/tool: fix query inode coredump
inodeId should be uint64_t, use std::stoi to convert the address out of
bounds, should be std::stoull

+++ curvefs/src/tools/query/curvefs_inode.cpp
@@ -86,11 +86,11 @@ int InodeTool::Init() {
 
     for (size_t i = 0; i < poolsId.size(); ++i) {
         curvefs::metaserver::GetInodeRequest request;
-        request.set_poolid(std::stoi(poolsId[i]));
-        request.set_copysetid(std::stoi(copysetsId[i]));
-        request.set_partitionid(std::stoi(partitionId[i]));
-        request.set_fsid(std::stoi(fsId[i]));
-        request.set_inodeid(std::stoi(inodeId[i]));
+        request.set_poolid(std::stoul((poolsId[i])));
+        request.set_copysetid(std::stoul((copysetsId[i])));
+        request.set_partitionid(std::stoul((partitionId[i])));
+        request.set_fsid(std::stoul((fsId[i])));
+        request.set_inodeid(std::stoull(inodeId[i]));
         SetStreamingRpc(false);
         request.set_supportstreaming(true);
         AddRequest(request);
